==> 06_api_access_development.ipynb <==
import pandas as pd
import yaml
df = pd.read_csv('./data/jira_data.csv')

# Extract values from the DataFrame
epiclink = df["epiclink"].iloc[0]
epicname = df["epicname"].iloc[0]
epicstatus = df["epicstatus"].iloc[0]
project_key = df["project_keys"].iloc[0]

# Load existing YAML template
template_file = "sample-config.yaml"
with open(template_file, "r") as file:
    config = yaml.safe_load(file)
# Update YAML values with variables from CSV
config["serverfrom"]["customfields"]["epiclink"] = epiclink
config["serverfrom"]["customfields"]["epicname"] = epicname
config["serverfrom"]["customfields"]["epicstatus"] = epicstatus
config["serverfrom"]["projectkey"] = project_key

# Save updated YAML
with open(template_file, "w") as file:
    yaml.dump(config, file, default_flow_style=False)

print(f"Updated {template_file} with values from CSV.")
df.head()
